fixed installer error

+++ scripts/install_grpcurl.sh
@@ -7,8 +7,8 @@ EXECUTABLE="grpcurl"
 TARBALL="${EXECUTABLE}_${VERSION}_linux_x86_64.tar.gz"
 URL="https://github.com/fullstorydev/grpcurl/releases/download/v${VERSION}/${TARBALL}"
 
-if [ $1 = "delete" ]; then
-  rm -f "${HOME}/bin/${EXECUTABLE}"
+if [ $# -eq 1 ] && [ $1 = "delete" ]; then
+    rm -f "${HOME}/bin/${EXECUTABLE}"
 fi
 
 if [ ! -f "${HOME}/bin/${EXECUTABLE}" ] ; then
